pyproject.toml
[tool.poetry]
name = "pylibsdm"
version = "0.1.0"
description = "Python library for handling Secure Dynamic Messaging (SDM) of NFC cards like the NTAG 424 DNA"
authors = ["Dominik George <nik@naturalnet.de>"]
repository = "https://codeberg.org/Natureshadow/pylibsdm"
license = "Apache-2.0"
keywords = ["nfc", "ntag424", "an12196"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.18.0"
nfcpy = "^1.0.4"
typer = { version = "^0.9.0", optional = true, extras = ["all"] }
pydantic = "^1.10.9"

[tool.poetry.extras]
cli = ["typer"]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.274"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.1"

[tool.poetry.scripts]
sdmutil = { callable = "pylibsdm.cli:app", extras = ["cli"] }

[tool.poetry.plugins."pylibsdm.tags"]
ntag424dna = "pylibsdm.tag.ntag424dna:Tag"

[tool.black]
line-length = 88
target-version = ['py311']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
